refactor(users): move passwordProblem out of the component file (CRM-210)
ESLint (diff) was red on the PR:

    Fast refresh only works when a file only exports components. Use a new file
    to share constants or functions between components
    react-refresh/only-export-components

I had exported passwordProblem from SetPasswordModal.tsx so the rule could be
tested against app/models/user.rb without driving the dialog. That breaks Fast
Refresh for the whole module.

Moved to passwordRules.ts, which is where it belonged anyway — a validation rule
mirroring Devise's password_length and User#password_complexity is not a
component concern. The test keeps importing it directly.

eslint clean on the touched files, tsc --noEmit clean, 37 tests green.
